Adding code to deal with ContentFilter errors
This change is associated with another change in the
Readium SDK. In the Readium SDK code, I created a new
architecture that lets ContentFilters report back to the app
when an error happens in their processing. This code, then,
will listen to those errors, and then the Launcher will log the
error when there is one.